add Photon events to synchronize room countdown timer across network.

+++ Assets/Examples/Config/Scripts/RuntimeGameConfig.cs
@@ -38,7 +38,7 @@ namespace Examples.Config.Scripts
     [Serializable]
     public class GameVariables
     {
-        [Header("Battle")] public float roomStartDelay;
+        [Header("Battle")] public int roomStartDelay;
 
         [Header("Ball")] public float ballMoveSpeed;
         public float ballLerpSmoothingFactor;
